Laboratorium7/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-widget-tweaks==1.4.8
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
pytz==2020.5
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
